Bugfixes in M17 implementation, bitstream matches
Fixed a bug in the callsing encoding and in the randomizer.
The callsign copy was going past the end of the callsign itself and the
randomizer was using an array allocated in the header file, which in
miosix shows up as 0xFFs. Now the bitstream matches with the linux
reference implementation with codec2 disabled.